style: post-detail answers show divider
