+++ spotcheck-ios/Components/Cells/Controls/VotingControls.swift
