--- spotcheck-ios/Components/FeedCell.swift
@@ -1,202 +1,0 @@
-import MaterialComponents
-import PromiseKit
-
-enum OverflowMenuLocation {
-    case top, bottom
-}
-
-class FeedCell: MDCCardCollectionCell {
-    static let IMAGE_HEIGHT = 200
-    
-    lazy var widthConstraint: NSLayoutConstraint = {
-        let width = contentView.widthAnchor.constraint(equalToConstant: bounds.size.width)
-        width.isActive = true
-        return width
-    }()    
-    
-    var postId: ExercisePostID?
-    var mediaHeightConstraint: NSLayoutConstraint?
-    var post: ExercisePost?
-    var dividerLeadingConstraint: NSLayoutConstraint?
-    var votingControls: VotingControls = {
-        let controls = VotingControls()
-        controls.translatesAutoresizingMaskIntoConstraints = false
-        return controls
-    }()
-    var overflowMenuTap: (() -> Void)? = nil
-    private var overflowMenuLayoutConstraints: [NSLayoutConstraint]?
-    
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-        contentView.translatesAutoresizingMaskIntoConstraints = false
-                
-        applyTheme(withScheme: ApplicationScheme.instance.containerScheme)
-        addSubviews()
-        applyConstraints()
-        initControls()
-    }
-    
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-    
-    override func systemLayoutSizeFitting(_ targetSize: CGSize, withHorizontalFittingPriority horizontalFittingPriority: UILayoutPriority, verticalFittingPriority: UILayoutPriority) -> CGSize {
-        widthConstraint.constant = bounds.size.width
-        return contentView.systemLayoutSizeFitting(CGSize(width: targetSize.width, height: 1))
-    }
-        
-    override func prepareForReuse() {
-        super.prepareForReuse()
-        //when cell is being reused, must reset every property since cell isn't fully cleaned automatically
-        media.sd_cancelCurrentImageLoad()
-        media.image = UIImage(named:"squatLogoPlaceholder")!//nil
-    }
-
-    func addSubviews() {
-        //contentView.addSubview(thumbnailImageView)
-        contentView.addSubview(headerLabel)
-        contentView.addSubview(subHeadLabel)
-        contentView.addSubview(media)
-        contentView.addSubview(supportingTextLabel)
-        contentView.addSubview(votingControls)
-        contentView.addSubview(overflowMenu)
-        contentView.addSubview(cellDivider)
-    }
-    
-    func applyConstraints() {
-        
-        mediaHeightConstraint = media.heightAnchor.constraint(equalToConstant: CGFloat(0))
-        dividerLeadingConstraint = cellDivider.leadingAnchor.constraint(equalTo: contentView.leadingAnchor, constant: 16)
-        
-        NSLayoutConstraint.activate([
-            
-//        thumbnailImageView.topAnchor.constraint(equalTo: contentView.topAnchor, constant: 16),
-//        thumbnailImageView.leadingAnchor.constraint(equalTo: contentView.leadingAnchor, constant: 16),
-//        thumbnailImageView.widthAnchor.constraint(equalToConstant: 40),
-//        thumbnailImageView.heightAnchor.constraint(equalToConstant: 40),
-
-        headerLabel.topAnchor.constraint(equalTo: contentView.topAnchor, constant: 16),
-        headerLabel.leadingAnchor.constraint(equalTo: contentView.leadingAnchor, constant: 16),
-        
-        subHeadLabel.topAnchor.constraint(equalTo: headerLabel.bottomAnchor, constant: 8),
-        subHeadLabel.leadingAnchor.constraint(equalTo: contentView.leadingAnchor, constant: 16),
-        
-        media.topAnchor.constraint(equalTo: subHeadLabel.bottomAnchor, constant: 16),
-        media.widthAnchor.constraint(equalToConstant: contentView.frame.width),
-        
-        supportingTextLabel.topAnchor.constraint(equalTo: media.bottomAnchor, constant: 16),
-        supportingTextLabel.leadingAnchor.constraint(equalTo: contentView.leadingAnchor, constant: 16),
-        supportingTextLabel.trailingAnchor.constraint(equalTo: contentView.trailingAnchor, constant: -16),
-        
-        votingControls.topAnchor.constraint(equalTo: supportingTextLabel.bottomAnchor, constant: 24),
-        dividerLeadingConstraint!,
-        cellDivider.topAnchor.constraint(equalTo: votingControls.bottomAnchor, constant: 8),
-        contentView.trailingAnchor.constraint(equalTo: cellDivider.trailingAnchor),
-        contentView.bottomAnchor.constraint(equalTo: cellDivider.bottomAnchor)
-        ])
-    }
-    
-    func initControls() {
-        overflowMenu.view.addTarget(self, action: #selector(overflowMenuOnTapped(_:)), for: .touchUpInside)
-    }
-    
-    func setConstraintsWithMedia() {
-        mediaHeightConstraint!.constant = CGFloat(FeedCell.IMAGE_HEIGHT)
-        mediaHeightConstraint!.isActive = true
-        media.isHidden = false
-    }
-    
-    func setConstraintsWithNoMedia() {
-        mediaHeightConstraint!.constant = 0
-        mediaHeightConstraint!.isActive = true
-        media.isHidden = true
-    }
-    
-    func setOverflowMenuLocation(location: OverflowMenuLocation) {
-        if overflowMenuLayoutConstraints == nil {
-            if location == .bottom {
-                overflowMenuLayoutConstraints = [overflowMenu.topAnchor.constraint(equalTo: supportingTextLabel.bottomAnchor, constant: 24),
-                contentView.trailingAnchor.constraint(equalTo: overflowMenu.trailingAnchor, constant: 8),
-                contentView.bottomAnchor.constraint(equalTo: overflowMenu.bottomAnchor, constant: 16)]
-            }
-            else {
-                overflowMenuLayoutConstraints = [overflowMenu.topAnchor.constraint(equalTo: contentView.topAnchor, constant: 8),
-                                                 contentView.trailingAnchor.constraint(equalTo: overflowMenu.trailingAnchor, constant: 8)]
-            }
-            
-            NSLayoutConstraint.activate(overflowMenuLayoutConstraints!)
-        }
-    }
-    
-    @objc func overflowMenuOnTapped(_ sender: Any) {
-        if let event = overflowMenuTap {
-            event()
-        }
-    }
-    
-    func hideDivider() {
-        cellDivider.isHidden = true
-    }
-    
-    func setFullBleedDivider() {
-        dividerLeadingConstraint?.constant = 0
-    }
-    
-    let headerLabel: UILabel = {
-        let label = UILabel()
-        label.translatesAutoresizingMaskIntoConstraints = false
-        label.font = ApplicationScheme.instance.containerScheme.typographyScheme.body2
-        label.textColor = ApplicationScheme.instance.containerScheme.colorScheme.onBackgroundColor
-        label.numberOfLines = 2
-        label.lineBreakMode = .byWordWrapping
-        return label
-    }()
-    
-    let subHeadLabel: UILabel = {
-        let label = UILabel()
-        label.translatesAutoresizingMaskIntoConstraints = false
-        label.font = ApplicationScheme.instance.containerScheme.typographyScheme.body2
-        label.textColor = ApplicationScheme.instance.containerScheme.colorScheme.primaryColorVariant
-        return label
-    }()
-    
-    let thumbnailImageView: UIImageView = {
-        let image = UIImageView()
-        image.translatesAutoresizingMaskIntoConstraints = false
-        image.image = Images.profilePictureDefault
-        image.tintColor = .white
-        //TODO: Clicking on the image should take you to its location.
-        return image
-    }()
-    
-    let media: UIImageView = {
-        let image = UIImageView()
-        image.translatesAutoresizingMaskIntoConstraints = false
-        image.contentMode = .scaleAspectFit
-        return image
-    }()
-    
-    let supportingTextLabel: UILabel = {
-        let label = UILabel()
-        label.translatesAutoresizingMaskIntoConstraints = false
-        label.font = ApplicationScheme.instance.containerScheme.typographyScheme.body2
-        label.textColor = ApplicationScheme.instance.containerScheme.colorScheme.onBackgroundColor
-        label.numberOfLines = 3
-        label.lineBreakMode = .byWordWrapping
-        return label
-    }()
-    
-    let overflowMenu: FlatButton = {
-        let button = FlatButton()
-        button.translatesAutoresizingMaskIntoConstraints = false
-        button.view.setImage(Images.moreHorizontal, for: .normal)
-        button.view.setImageTintColor(ApplicationScheme.instance.containerScheme.colorScheme.onSurfaceColor, for: .normal)
-        return button
-    }()
-    
-    let cellDivider: Divider = {
-        let divider = Divider()
-        divider.translatesAutoresizingMaskIntoConstraints = false
-        return divider
-    }()
-}
